+++ moabb/pipelines/features.py
@@ -1,6 +1,7 @@
 import numpy as np
 import scipy.signal as signal
 from sklearn.base import BaseEstimator, TransformerMixin
+from sklearn.preprocessing import StandardScaler
 
 
 class LogVariance(BaseEstimator, TransformerMixin):
@@ -61,3 +62,61 @@ class ExtendedSSVEPSignal(BaseEstimator, TransformerMixin):
         n_trials, n_freqs, n_channels, n_times = out.shape
         out = out.reshape((n_trials, n_channels * n_freqs, n_times))
         return out
+
+
+class AugmentedDataset(BaseEstimator, TransformerMixin):
+    """This transformation allow to create an embedding version of the current dataset.
+    The implementation and the application is described in [1]_.
+
+    References
+    ----------
+    .. [1] Carrara, I., & Papadopoulo, T. (2023). Classification of BCI-EEG based on augmented covariance matrix.
+           arXiv preprint arXiv:2302.04508.
+           https://doi.org/10.48550/arXiv.2302.04508
+    """
+
+    def __init__(self, order=1, lag=1):
+        self.order = order
+        self.lag = lag
+
+    def fit(self, X, y):
+        return self
+
+    def transform(self, X):
+        if self.order == 1:
+            X_fin = X
+        else:
+            X_fin = []
+
+            for i in np.arange(X.shape[0]):
+                X_p = X[i][:, : -self.order * self.lag]
+                for p in np.arange(1, self.order):
+                    X_p = np.append(
+                        X_p, X[i][:, p * self.lag : -(self.order - p) * self.lag], axis=0
+                    )
+                X_fin.append(X_p)
+            X_fin = np.array(X_fin)
+
+        return X_fin
+
+
+class StandardScaler_Epoch(BaseEstimator, TransformerMixin):
+    """
+    Function to standardize the X raw data for the DeepLearning Method
+    """
+
+    def __init__(self):
+        """Init."""
+
+    def fit(self, X, y):
+        return self
+
+    def transform(self, X):
+        X_fin = []
+
+        for i in np.arange(X.shape[0]):
+            X_p = StandardScaler().fit_transform(X[i])
+            X_fin.append(X_p)
+        X_fin = np.array(X_fin)
+
+        return X_fin