+++ moabb/pipelines/utils.py
@@ -20,7 +20,9 @@ log = logging.getLogger(__name__)
 def create_pipeline_from_config(config):
     """Create a pipeline from a config file.
 
-    takes a config dict as input and return the coresponding pipeline.
+    takes a config dict as input and return the corresponding pipeline.
+
+    If the pipeline is a Tensorflow pipeline it convert also the optimizer function and the callbacks.
 
     Parameters
     ----------
@@ -41,6 +43,26 @@ def create_pipeline_from_config(config):
         # create the instance
         if "parameters" in component.keys():
             params = component["parameters"]
+            if "optimizer" in component["parameters"].keys():
+                for optm in component["parameters"]["optimizer"]:
+                    mod_optm = __import__(name=optm["from"], fromlist=[optm["name"]])
+                    params_optm = optm["parameters"]
+                    instance = getattr(mod_optm, optm["name"])(**params_optm)
+                    component["parameters"]["optimizer"] = instance
+
+            if "callbacks" in component["parameters"].keys():
+                cb = []
+                for callbacks in component["parameters"]["callbacks"]:
+                    mod_callbacks = __import__(
+                        name=callbacks["from"], fromlist=[callbacks["name"]]
+                    )
+                    params_callbacks = callbacks["parameters"]
+                    instance = getattr(mod_callbacks, callbacks["name"])(
+                        **params_callbacks
+                    )
+                    cb.append(instance)
+                component["parameters"]["callbacks"] = cb
+
         else:
             params = {}
         instance = getattr(mod, component["name"])(**params)