+++ docs/source/whats_new.rst
@@ -22,6 +22,8 @@ Enhancements
 - Adding a comprehensive benchmarking function (:gh:`264` by `Divyesh Narayanan`_ and `Sylvain Chevallier`_)
 - Add meta-information for datasets in documentation (:gh:`317` by `Bruno Aristimunha`_)
 - Add GridSearchCV for different evaluation procedure (:gh:`319` by `Igor Carrara`_)
+- Add six deep learning models (Tensorflow), and build a tutorial to show to used the deep learning model (:gh:`326` by `Igor Carrara`_, `Bruno Aristimunha`_ and `Sylvain Chevallier`_)
+- Add a augmentation model to the pipeline (:gh:`326` by `Igor Carrara`_)
 
 Bugs
 ~~~~
@@ -32,6 +34,7 @@ Bugs
 - Correct usage of name simplification function in analyze (:gh:`306` by `Divyesh Narayanan`_)
 - Fix downloading path issue for Weibo2014 and Zhou2016, numy error in DemonsP300 (:gh:`315` by `Sylvain Chevallier`_)
 - Fix unzip error for Huebner2017 and Huebner2018 (:gh:`318` by `Sylvain Chevallier`_)
+- Fix n_classes when events set to None (:gh:`337` by `Igor Carrara`_ and `Sylvain Chevallier`_)
 
 API changes
 ~~~~~~~~~~~